Rest full V5

+++ src/main/java/org/bayat/crud/model/repository/DataRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface DataRepository extends JpaRepository<Data, Integer> {
+public interface DataRepository extends JpaRepository<Data, Long> {
     Optional<Data> findByPhone(String phone);
